Major Revision: Heading toward cloud-sourced use by non-expert voters More material ui look (without actually adding all of the material ui librarys) New system pop-up Off of the WeVote icon near the omnibar, To provide enabling options and some direct links to other pages. Simplified the right click pop up in the endorsement page to have only one choice. Disabled Links On the endorsement page HTML so you don't accidentally navigate away when you click it to edit endorsement.  Capture the disabled link href so we can reuse it in the dialog. Simplified the text in the sign in pop up.
resolves wevote/EndorsementExtension/issues/55   If the website already had the candidate name linked to another page, override that link and always open our popover
resolves wevote/EndorsementExtension/issues/54   For all candidate name highlights, open a popover when you click on the highlighted name.
resolves wevote/EndorsementExtension/issues/48 Add a pop-up menu for the 'W' extension icon after a Left Click.
resolves wevote/EndorsementExtension/issues/49  The icons now display in the DOM of the endorsement page.

February 18, 2020
Sign in is working well again, and it removes the dialogue and the inner panel.
The new way of getting the voter device ID is more robust and will fit well with the plans for a new dedicated sign in page in the web app. All we need the web app to do is change the active tab to the tab we came from, and then the extension can remove the dialogue and refresh the avatar.
Highlighting the current tab works perfectly, but slowly. Opening the editor works perfectly, but doesn't leave the endorsement iframe highlighted.
After getting highlighting in the I-frame running again, I want to finish adding highlighting on the chosen subset of tabs and the same for editor, and this section of design will be stable.